+++ src/fabric_cicd/changelog.md
@@ -6,6 +6,32 @@ The following contains all major, minor, and patch version release notes.
 -   📝 Documentation Update
 -   ⚡ Internal Optimization
 
+## Version 0.1.24
+
+<span class="md-h2-subheader">Release Date: 2025-08-04</span>
+
+-   💥 Require parameterization for Dataflow and Semantic Model references in Data Pipeline activities
+-   💥 Require specific parameterization for deploying a Dataflow that depends on another in the same workspace (see Parameterization docs)
+-   🔧 Fix Dataflow/Data Pipeline deployment failures caused by workspace permissions ([#419](https://github.com/microsoft/fabric-cicd/issues/419))
+-   🔧 Prevent duplicate logical ID issue in Report and Semantic Model deployment ([#405](https://github.com/microsoft/fabric-cicd/issues/405))
+-   🔧 Fix deployment of items without assigned capacity ([#402](https://github.com/microsoft/fabric-cicd/issues/402))
+-   📝 Improve Parameterization documentation ([#415](https://github.com/microsoft/fabric-cicd/issues/415))
+-   ⚡ Support for Eventhouse query URI parameterization ([#414](https://github.com/microsoft/fabric-cicd/issues/414))
+-   ⚡ Support for Warehouse SQL endpoint parameterization ([#392](https://github.com/microsoft/fabric-cicd/issues/392))
+
+## Version 0.1.23
+
+<span class="md-h2-subheader">Release Date: 2025-07-08</span>
+
+-   ✨ New functionalities for GitHub Copilot Agent and PR-to-Issue linking
+-   🔧 Fix issue with lakehouse shortcuts publishing ([#379] (https://github.com/microsoft/fabric-cicd/issues/379))
+-   🔧 Add validation for empty logical IDs to prevent deployment corruption ([#86](https://github.com/microsoft/fabric-cicd/issues/86))
+-   🔧 Fix SQL provision print statement ([#329](https://github.com/microsoft/fabric-cicd/issues/329))
+-   🔧 Rename the error code for reserved item name per updated Microsoft Fabric API ([#388](https://github.com/microsoft/fabric-cicd/issues/388))
+-   🔧 Fix lakehouse exclude_regex to exclude shortcut publishing ([#385](https://github.com/microsoft/fabric-cicd/issues/385))
+-   🔧 Remove max retry limit to handle large deployments ([#299] (https://github.com/microsoft/fabric-cicd/issues/299))
+-   📝 Fix formatting and examples in the How to and Examples pages
+
 ## Version 0.1.22
 
 <span class="md-h2-subheader">Release Date: 2025-06-25</span>